extracted helper functions and added jsdoc
